+++ packages/plugins/news/index.ts
@@ -1,2 +1,3 @@
 export * from './harness-tool.ts'
 export * from './plugin.ts'
+export * from './acquisition/index.ts'